requirements.txt
nbformat
matplotlib

torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0

pillow
num2words
faiss-cpu

unsloth
huggingface-hub
wandb

textdistance
sacrebleu

python-dotenv